feat: rewrite sd-radio-group documentation
