+++ packages/components/src/components/radio-group/radio-group.ts
@@ -64,6 +64,20 @@ export default class SdRadioGroup extends SolidElement implements SolidFormContr
    */
   @state() showInvalidStyle = false;
 
+  /** The radio group's size. This size will be applied to the label, all child radios and radio buttons. */
+  @property({ reflect: true }) size: 'lg' | 'sm' = 'lg';
+
+  /** Ensures a child radio is checked before allowing the containing form to submit. */
+  @property({ type: Boolean, reflect: true }) required = false;
+
+  /**
+   * The orientation property determines the alignment of the component's content or elements. It accepts two possible
+   * values: 'horizontal' and 'vertical'. The default value is 'vertical'.
+   * This property allows you to control the visual layout and arrangement of elements within the component, providing
+   * flexibility in how the component is displayed based on your specific design needs.
+   */
+  @property({ reflect: true }) orientation: 'horizontal' | 'vertical' = 'vertical';
+
   /**
    * The radio group's label. Required for proper accessibility. If you need to display HTML, use the `label` slot
    * instead.
@@ -82,26 +96,12 @@ export default class SdRadioGroup extends SolidElement implements SolidFormContr
   /** The current value of the radio group, submitted as a name/value pair with form data. */
   @property({ reflect: true }) value = '';
 
-  /** The radio group's size. This size will be applied to the label, all child radios and radio buttons. */
-  @property({ reflect: true }) size: 'lg' | 'sm' = 'lg';
-
   /**
    * By default, form controls are associated with the nearest containing `<form>` element. This attribute allows you
    * to place the form control outside of a form and associate it with the form that has this `id`. The form must be in
    * the same document or shadow root for this to work.
    */
   @property({ reflect: true }) form = '';
-
-  /** Ensures a child radio is checked before allowing the containing form to submit. */
-  @property({ type: Boolean, reflect: true }) required = false;
-
-  /**
-   * The orientation property determines the alignment of the component's content or elements. It accepts two possible
-   * values: 'horizontal' and 'vertical'. The default value is 'vertical'.
-   * This property allows you to control the visual layout and arrangement of elements within the component, providing
-   * flexibility in how the component is displayed based on your specific design needs.
-   */
-  @property({ reflect: true }) orientation: 'horizontal' | 'vertical' = 'vertical';
 
   /** Gets the validity state object */
   get validity() {
